ifcbapi/requirements.txt
Django==5.0.6
django-ninja==1.1.0
psycopg2==2.9.9
